Part 066

+++ lib/helpers/api_helper.dart
@@ -5,19 +5,24 @@ import 'package:vehicles_app/models/brand.dart';
 import 'package:vehicles_app/models/document_type.dart';
 import 'package:vehicles_app/models/procedure.dart';
 import 'package:vehicles_app/models/response.dart';
+import 'package:vehicles_app/models/token.dart';
 import 'package:vehicles_app/models/user.dart';
 import 'package:vehicles_app/models/vehicle_type.dart';
 import 'constans.dart';
 
 class ApiHelper {
-  static Future<Response> getProcedures(String token) async {
+  static Future<Response> getProcedures(Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+
     var url = Uri.parse('${Constans.apiUrl}/api/Procedures');
     var response = await http.get(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -37,14 +42,18 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> getBrands(String token) async {
+  static Future<Response> getBrands(Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+
     var url = Uri.parse('${Constans.apiUrl}/api/Brands');
     var response = await http.get(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -64,14 +73,18 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> getDocumentTypes(String token) async {
+  static Future<Response> getDocumentTypes(Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}/api/DocumentTypes');
     var response = await http.get(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -91,14 +104,18 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> getVehicleTypes(String token) async {
+  static Future<Response> getVehicleTypes(Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}/api/VehicleTypes');
     var response = await http.get(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -118,14 +135,18 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> getUsers(String token) async {
+  static Future<Response> getUsers(Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}/api/Users');
     var response = await http.get(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -145,14 +166,18 @@ class ApiHelper {
     return Response(isSuccess: true, result: list);
   }
 
-  static Future<Response> put(String controller, String id, Map<String, dynamic> request, String token) async {
+  static Future<Response> put(String controller, String id, Map<String, dynamic> request, Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}$controller$id');
     var response = await http.put(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
       body: jsonEncode(request),
     );
@@ -164,14 +189,18 @@ class ApiHelper {
     return Response(isSuccess: true);
   }
 
-  static Future<Response> post(String controller, Map<String, dynamic> request, String token) async {
+  static Future<Response> post(String controller, Map<String, dynamic> request, Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}$controller');
     var response = await http.post(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
       body: jsonEncode(request),
     );
@@ -183,14 +212,18 @@ class ApiHelper {
     return Response(isSuccess: true);
   }
 
-  static Future<Response> delete(String controller, String id, String token) async {
+  static Future<Response> delete(String controller, String id, Token token) async {
+    if (!_validToken(token)) {
+      return Response(isSuccess: false, message: 'Sus credenciales se han vencido, por favor cierre sesión y vuelva a ingresar al sistema.');
+    }
+    
     var url = Uri.parse('${Constans.apiUrl}$controller$id');
     var response = await http.delete(
       url,
       headers: {
         'content-type' : 'application/json',
         'accept' : 'application/json',
-        'authorization': 'bearer $token',
+        'authorization': 'bearer ${token.token}',
       },
     );
 
@@ -199,5 +232,13 @@ class ApiHelper {
     }
 
     return Response(isSuccess: true);
+  }
+
+  static bool _validToken(Token token) {
+    if (DateTime.parse(token.expiration).isAfter(DateTime.now())) {
+      return true;
+    }
+
+    return false;
   }
 }
